# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talai middleware",
    "description": "The only component allowed to talk to TallyPrime. Every route under /api/v1 requires a bearer token; /health does not.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/drafts": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Create a draft",
        "operationId": "create_draft_api_v1_drafts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPurchaseBill-Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "List drafts",
        "operationId": "list_drafts_api_v1_drafts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "One draft",
        "operationId": "get_draft_api_v1_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "drafts"
        ],
        "summary": "Replace a draft",
        "operationId": "update_draft_api_v1_drafts__draft_id__put",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPurchaseBill-Input"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "drafts"
        ],
        "summary": "Delete a draft",
        "operationId": "delete_draft_api_v1_drafts__draft_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/validate": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Run the rules",
        "operationId": "validate_draft_api_v1_drafts__draft_id__validate_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drafts/{draft_id}/queue": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Queue for sync",
        "operationId": "queue_draft_api_v1_drafts__draft_id__queue_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attachments": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Upload a bill file",
        "description": "Stores the file under ``UPLOAD_DIR`` with a random name (plan \u00a76).\n\nOCR is mocked in v1: the attachment is recorded with ``ocr_status='skipped'``.",
        "operationId": "upload_attachment_api_v1_attachments_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_attachment_api_v1_attachments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/pull": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Pull from Tally now",
        "operationId": "pull_api_v1_sync_pull_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Commit queued drafts",
        "operationId": "push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/runs": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Recent sync runs",
        "operationId": "runs_api_v1_sync_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/runs/{run_id}": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "One sync run",
        "operationId": "run_api_v1_sync_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttachmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "ocr_status": {
            "type": "string",
            "title": "Ocr Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "draft_id",
          "file_name",
          "mime",
          "size_bytes",
          "ocr_status",
          "created_at"
        ],
        "title": "AttachmentOut"
      },
...
      "Body_upload_attachment_api_v1_attachments_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_attachment_api_v1_attachments_post"
      },
...
      "DraftList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DraftOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "DraftList"
      },
      "DraftOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "validated",
              "queued",
              "committing",
              "committed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DraftPurchaseBill-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Errors"
          },
          "generated_xml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Xml"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          },
          "tally_voucher_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tally Voucher Number"
          },
          "tally_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tally Guid"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DraftOut"
      },
      "DraftPurchaseBill-Input": {
        "properties": {
          "gst_registration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst Registration"
          },
          "voucher_type": {
            "type": "string",
            "title": "Voucher Type",
            "default": "Purchase"
          },
          "voucher_date": {
            "type": "string",
            "format": "date",
            "title": "Voucher Date"
          },
          "bill_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bill Date"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "supplier_invoice_no": {
            "type": "string",
            "title": "Supplier Invoice No"
          },
          "cost_centre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Centre"
          },
          "party": {
            "$ref": "#/components/schemas/PartyPayload"
          },
          "purchase_ledger": {
            "type": "string",
            "title": "Purchase Ledger"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemPayload-Input"
            },
            "type": "array",
            "title": "Items"
          },
          "ledger_lines": {
            "items": {
              "$ref": "#/components/schemas/LedgerLinePayload-Input"
            },
            "type": "array",
            "title": "Ledger Lines"
          },
          "tax_lines": {
            "items": {
              "$ref": "#/components/schemas/LedgerLinePayload-Input"
            },
            "type": "array",
            "title": "Tax Lines"
          },
          "reverse_charge": {
            "type": "boolean",
            "title": "Reverse Charge",
            "default": false
          },
          "narration": {
            "type": "string",
            "title": "Narration",
            "default": ""
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsPayload-Input"
          },
          "allow_duplicate": {
            "type": "boolean",
            "title": "Allow Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "voucher_date",
          "supplier_invoice_no",
          "party",
          "purchase_ledger"
        ],
        "title": "DraftPurchaseBill",
        "description": "The Create Bill form payload (plan \u00a73.6)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartyPayload": {
        "properties": {
          "ledger_name": {
            "type": "string",
            "title": "Ledger Name"
          },
          "gstin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gstin"
          },
          "gst_treatment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gst Treatment"
          },
          "billing_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Address"
          },
          "source_of_supply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Of Supply"
          },
          "destination_of_supply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Of Supply"
          }
        },
        "type": "object",
        "required": [
          "ledger_name"
        ],
        "title": "PartyPayload"
      },
      "PullRequest": {
        "properties": {
          "scopes": {
            "items": {
              "type": "string",
              "enum": [
                "masters",
                "vouchers",
                "bills"
              ]
            },
            "type": "array",
            "title": "Scopes"
          },
          "from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Date"
          },
          "to_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Date"
          }
        },
        "type": "object",
        "title": "PullRequest"
      },
      "PushRequest": {
        "properties": {
          "draft_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Ids"
          }
        },
        "type": "object",
        "title": "PushRequest"
      },
      "PushResponse": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/SyncRunOut"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PushResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "run",
          "results"
        ],
        "title": "PushResponse"
      },
...
      "SyncRunList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SyncRunOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SyncRunList"
      },
      "SyncRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "records_seen": {
            "type": "integer",
            "title": "Records Seen"
          },
          "records_changed": {
            "type": "integer",
            "title": "Records Changed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "scope",
          "status",
          "started_at",
          "records_seen",
          "records_changed"
        ],
        "title": "SyncRunOut"
      },
...
      "ValidationResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "validated",
              "queued",
              "committing",
              "committed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "status",
          "errors"
        ],
        "title": "ValidationResult"
      },
...
    }
  }
}